example_django_rest_api/requirement.txt
argh==0.26.2
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.4
django-livesync==0.4
django-rest-swagger==2.2.0
djangorestframework==3.10.2
drf-yasg==1.16.1
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==19.1
pathtools==0.1.2
pyparsing==2.4.2
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
ruamel.yaml==0.16.0
ruamel.yaml.clib==0.1.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
tornado==4.5.1
uritemplate==3.0.0
urllib3==1.25.3
uWSGI==2.0.18
watchdog==0.9.0
websocket-client==0.56.0
